+++ tests/heap_lang.v
@@ -366,7 +366,17 @@ Section tests.
   Proof.
     iIntros "Hcred".
     wp_pure credit:"Hcred". Show.
-    iIntros "!> HP". iNext credit: "Hcred". auto.
+    iIntros "!> HP". iNext credit: "Hcred".
+    Show. (* Check that the order of hypotheses is preserved *) auto.
+  Qed.
+
+  Check "test_wp_pure_credit_bump_2".
+  Lemma test_wp_pure_credit_bump_2 P :
+    ⊢ £ 1 -∗ WP #42 + #420 {{ v, ▷ ▷ P ={∅}=∗ P }}.
+  Proof.
+    iIntros "Hcred".
+    wp_pure credit:"Hcred". Show.
+    iIntros "!> HP". iNext 2 credit: "Hcred". auto.
   Qed.
 
   Check "test_wp_pure_credit_fail".